requirements.txt
setuptools~=72.1.0
streamlit==1.37.0
streamlit-option-menu==0.3.13
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
langchain~=0.2.15
langchain-community~=0.2.14
langchain-openai~=0.1.23
langchain-chroma~=0.1.3
langchain-huggingface~=0.0.3
pypdf==4.3.1